feat(fix): Fix error on build Next app, moving style.js from page folder
